v.2.0_first_commit_after_review_Mesto3

+++ src/components/api.js
@@ -6,20 +6,29 @@ const getResponse = (res) => {
     }
 };
 
+const config = {
+    baseUrl: 'https://nomoreparties.co/v1/wff-cohort-37',
+    headers: {
+        authorization: "1efcbbe3-d8c4-4422-be61-93be6ca1a8ea",
+        
+    },
+    headersCT: {
+        authorization: "1efcbbe3-d8c4-4422-be61-93be6ca1a8ea",
+        'Content-Type': 'application/json'
+    }
+};
+
+
 const personalPromis = () => {
-    return fetch("https://nomoreparties.co/v1/wff-cohort-37/users/me", {
-        headers: {
-            authorization: "1efcbbe3-d8c4-4422-be61-93be6ca1a8ea",
-        },
+    return fetch(`${config.baseUrl}/users/me`, {
+        headers: config.headers
     })
     .then(getResponse);
 };
 
 const loadCardsPromis = () => {
-    return fetch("https://nomoreparties.co/v1/wff-cohort-37/cards", {
-        headers: {
-            authorization: "1efcbbe3-d8c4-4422-be61-93be6ca1a8ea",
-        },
+    return fetch(`${config.baseUrl}/cards`, {
+        headers: config.headers
     })
     .then(getResponse)
 };
@@ -27,27 +36,21 @@ const loadCardsPromis = () => {
 const promises = [personalPromis(), loadCardsPromis()];
 
 const safeServerEditProfile = (profileTitle, profileDescription) => {
-    return fetch("https://nomoreparties.co/v1/wff-cohort-37/users/me", {
+    return fetch(`${config.baseUrl}/users/me`, {
         method: 'PATCH',
-        headers: {
-            authorization: "1efcbbe3-d8c4-4422-be61-93be6ca1a8ea",
-            'Content-Type': 'application/json'
-        },
+        headers: config.headersCT,
         body: JSON.stringify({
-            name: profileTitle.textContent,
-            about: profileDescription.textContent,
+            name: profileTitle,
+            about: profileDescription,
         }) 
     })
     .then(getResponse)
 };
 
 const safeServerNewProfileAvatar = (linkImage) => {
-    return fetch("https://nomoreparties.co/v1/wff-cohort-37/users/me/avatar", {
+    return fetch(`${config.baseUrl}/users/me/avatar`, {
         method: 'PATCH',
-        headers: {
-            authorization: "1efcbbe3-d8c4-4422-be61-93be6ca1a8ea",
-            'Content-Type': 'application/json'
-        },
+        headers: config.headersCT,
         body: JSON.stringify({
             avatar: linkImage
         }) 
@@ -56,12 +59,9 @@ const safeServerNewProfileAvatar = (linkImage) => {
 };
 
 const addNewCardToServer = (newCard) => {
-    return fetch("https://nomoreparties.co/v1/wff-cohort-37/cards", {
+    return fetch(`${config.baseUrl}/cards`, {
         method: 'POST',
-        headers: {
-            authorization: "1efcbbe3-d8c4-4422-be61-93be6ca1a8ea",
-            'Content-Type': 'application/json'
-        },
+        headers: config.headersCT,
         body: JSON.stringify({
             name: newCard.name,
             link: newCard.link,
@@ -71,33 +71,27 @@ const addNewCardToServer = (newCard) => {
 };
 
 const deleteCardFromServer = (cardId) => {
-    return fetch(`https://nomoreparties.co/v1/wff-cohort-37/cards/${cardId}`, {
+    return fetch(`${config.baseUrl}/cards/${cardId}`, {
         method: 'DELETE',
-        headers: {
-            authorization: "1efcbbe3-d8c4-4422-be61-93be6ca1a8ea",
-        }
+        headers: config.headers
     })
     .then(getResponse)
   };
   
-  const likeCardOnServer = (cardId) => {
-    return fetch(`https://nomoreparties.co/v1/wff-cohort-37/cards/likes/${cardId}`, {
-      method: 'PUT',
-      headers: {
-          authorization: "1efcbbe3-d8c4-4422-be61-93be6ca1a8ea",
-      }
-    })
-    .then(getResponse)
-  };
-  
-  const deleteLikeCardOnServer = (cardId) => {
-    return fetch(`https://nomoreparties.co/v1/wff-cohort-37/cards/likes/${cardId}`, {
-      method: 'DELETE',
-      headers: {
-          authorization: "1efcbbe3-d8c4-4422-be61-93be6ca1a8ea",
-      }
-    })
-    .then(getResponse)
-  };
-  
-export { getResponse, safeServerEditProfile, safeServerNewProfileAvatar, addNewCardToServer, deleteCardFromServer, likeCardOnServer, deleteLikeCardOnServer, promises };
+  const setLikeOnServer = (cardId, isLiked) => {
+    if (!isLiked) {
+        return fetch(`${config.baseUrl}/cards/likes/${cardId}`, {
+            method: 'PUT',
+            headers: config.headers
+          })
+          .then(getResponse)
+    } else {
+        return fetch(`${config.baseUrl}/cards/likes/${cardId}`, {
+            method: 'DELETE',
+            headers: config.headers
+          })
+          .then(getResponse)
+    }
+  }
+ 
+export { getResponse, safeServerEditProfile, safeServerNewProfileAvatar, addNewCardToServer, deleteCardFromServer, setLikeOnServer, promises };
